LPS-150185 Move logic not depending on org.apache.http to HttpHelperUtil， add try catch
